requirements.txt
# Core Web Frameworks
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# Database & Storage
psycopg2-binary>=2.9.9
supabase>=2.0.0
python-dotenv>=1.0.0

# NLP & ML Libraries
transformers>=4.35.0
torch>=2.0.0
nltk>=3.8.0
spacy>=3.7.0
scikit-learn>=1.3.0
rake-nltk>=1.0.6

# Audio Processing (Optional)
# openai-whisper>=20231117
soundfile>=0.12.0
pydub>=0.25.0

# Data Visualization
plotly>=5.18.0
matplotlib>=3.8.0
wordcloud>=1.9.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
